Maintain an association between the original item used to create a catalog listing so that the creator can update the listing, which will preserve its purchase history, tags, rating, etc. In the future, we can potentially let owners of the old item upgrade to the new item or some other complicated fiddly business.

+++ src/java/com/threerings/msoy/item/server/ItemManager.java
@@ -857,6 +857,9 @@ public class ItemManager
                 if (item.used != 0) {
                     throw new InvocationException(ItemCodes.E_ITEM_IN_USE);
                 }
+                if (item.catalogId != 0) {
+                    throw new InvocationException(ItemCodes.E_ITEM_LISTED);
+                }
                 repo.deleteItem(ident.itemId);
                 return null;
             }
